pyaiot/gateway/coap: reorganize files and use new base gateway class
